Fix data type of time delta (No wonder the error was always 0).

+++ src/amr/error-estimation.h
@@ -11,7 +11,7 @@ namespace ErrorEstimation {
 template <class ProblemType, class Fields, unsigned padding>
 std::optional<CellFlags::Flags>
 getFlags(const PaddedGrid<Fields, padding>& grid, const std::array<unsigned, 3> gridOffset,
-         const unsigned timeDelta, const double errorThreshold,
+         const double timeDelta, const double errorThreshold,
          const Problem<ProblemType, Fields, padding>& problem,
          const std::array<BoundaryType, Faces::FaceMax>& boundaryTypes);
 };
@@ -95,7 +95,7 @@ inline bool checkThreshold(const PaddedGrid<Fields, padding>& upper,
 template <class ProblemType, class Fields, unsigned padding>
 std::optional<CellFlags::Flags>
 ErrorEstimation::getFlags(const PaddedGrid<Fields, padding>& grid,
-                          const std::array<unsigned, 3> gridOffset, const unsigned timeDelta,
+                          const std::array<unsigned, 3> gridOffset, const double timeDelta,
                           const double errorThreshold,
                           const Problem<ProblemType, Fields, padding>& problem,
                           const std::array<BoundaryType, Faces::FaceMax>& boundaryTypes) {
